requirements.txt
selenium==4.8.3
elasticsearch==8.6.2
pymongo==4.3.3
PyYAML==6.0